+++ docs/changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v1.4.0
+
+In progress.
+
+Added:
+
+* Decayed photons can be included in the HDF5 output file, if desired, via new optional flag `dump_decayed_photons`.
+* Tracking of laser depletion: new output variable `absorption` can be used to examine how much energy electrons and positrons have drawn from the laser pulse.
+
 ## v1.3.5
 
 2024-03-08